Refactor : Introduced Optimistic locking in client
 I just introduced the optimistic locking in the app where I release the lock in the changeProtocol state method of reader thread instead of doing in the run method of main thread. Now, the client doesn't freeze if I tap too rapidly.

Freeze Reason : main thread was not releasing the lock in a condition where lineUp/lineDown tap Is generated by user. When main thread tried to acquire the lock so it can send a message/change the state, but reader thread was changing the state. So, main thread was in blocked state and main thread had to release the lock in run() method which it couldn't do. Hence the deadlock.